refactor: 커스텀 훅 생성
