+++ src/App.jsx
@@ -1,30 +1,14 @@
-import { useState, useEffect } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Header from "./components/Header";
 import HomePage from "./pages/HomePage";
 import ToDoList from "./pages/ToDoList";
 import Login from "./pages/Login";
 import Signup from "./pages/Signup";
+import useAuth from "./hooks/useAuth";
 
 function App() {
-  const [user, setUser] = useState(null);
+  const { user, handleLogin, handleLogout } = useAuth();
 
-  useEffect(() => {
-    const storedUser = localStorage.getItem("user");
-    if (storedUser) {
-      setUser(JSON.parse(storedUser));
-    }
-  }, []);
-
-  const handleLogin = (userData) => {
-    setUser(userData);
-    localStorage.setItem("user", JSON.stringify(userData));
-  };
-
-  const handleLogout = () => {
-    setUser(null);
-    localStorage.removeItem("user");
-  };
   return (
     <BrowserRouter>
       <Header user={user} handleLogout={handleLogout} />
